Codesign: Initial implementation of macOS code signer
Main goal of this change is to make codesign step of release signing
to work on macOS. Development is currently was done on Linux based on
looking into what commands are being invoked.

Part of the change which touches other code signing utilities is
related on making it possible to test "foreign" code signers on a
different platform.